app/requirements.txt
Django==1.10.3
djangorestframework==3.5.3
djangorestframework-jwt==1.8.0
mysqlclient==1.3.9
py-auth0-jwt==0.2.4
py-auth0-jwt-rest==0.1
PyJWT==1.4.2
requests==2.11.1
